enable testing for server code
