+++ tools/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "app",
+  "version": "0.0.0",
+  "main": "./app.js"
+}
